refactor: Provide default implementations
